pyproject.toml
[project]
name = "rag-langchain-groq"
version = "0.1.0"
description = "RAG pipeline using LangChain, ChromaDB, and Groq LLMs"
authors = [
  { name = "Antonio Fiscarelli" }
]
readme = "README.md"
requires-python = ">=3.13.2"

dependencies = [
  "ipykernel",
  "langchain",  # meta package, optional if using modular packages below
  "langchain-core",
  "langchain-community",
  "langchain-text-splitters",
  "langchain-groq",
  "sentence-transformers",
  "chromadb",
  "python-dotenv",
  "pypdf",
  "pymupdf"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
